# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PilingTrack API",
    "description": "Industrial pile driving management platform — REST API for operators, dispatchers, and administrators.",
    "version": "2.0.0",
    "contact": {
      "name": "PilingTrack",
      "email": "support@pilingtrack.local"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    },
    {
      "url": "https://pilingtrack.example.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "pt-session",
        "description": "Session cookie (set via /api/auth/login)"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token (alternative to cookie)"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "DISPATCHER",
              "OPERATOR",
              "ASSISTANT"
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "New token pair"
          },
          "401": {
            "description": "Invalid refresh token"
          }
        }
      },
      "delete": {
        "summary": "Revoke refresh token",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Token revoked"
          }
        }
      }
    },
...
    "/api/reports/edit": {
      "get": {
        "summary": "Get editable report",
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "siteId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report data"
          }
        }
      }
    },
...
  }
}